Support for async flops
Adds limited support for asynchronous flops (as synchronous)

Corrects assupmtions processing

Changes to API

+++ avr.py
@@ -19,21 +19,23 @@ DEFAULT_CLK="clk"
 DEFAULT_TIMEOUT=3600
 DEFAULT_MEMOUT=4096
 DEFAULT_SPLIT=True
+DEFAULT_GRANULARITY=0
 DEFAULT_RANDOM=False
 DEFAULT_VERBOSITY=0
 	
 def getopts(header):
 	p = argparse.ArgumentParser(description=str(header), formatter_class=argparse.RawDescriptionHelpFormatter)
-	p.add_argument('-f', '--file', required=True, help='top file name', type=str)
+	p.add_argument('file', help='top file name', type=str)
 	p.add_argument('-t', '--top',       help='top module name (default: autodetect)', type=str)
 	p.add_argument('-n', '--name',      help='<test-name> (default: %s)' % DEFAULT_NAME, type=str, default=DEFAULT_NAME)
 	p.add_argument('-o', '--out',       help='<output-path> (default: %s)' % DEFAULT_OUT, type=str, default=DEFAULT_OUT)
 	p.add_argument('-b', '--bin',       help='binary path (default: %s)' % DEFAULT_BIN, type=str, default=DEFAULT_BIN)
-	p.add_argument('-y', '--yosys',           help='path to yosys installation (default: %s)' % DEFAULT_YOSYS, type=str, default=DEFAULT_YOSYS)
+	p.add_argument('-y', '--yosys',     help='path to yosys installation (default: %s)' % DEFAULT_YOSYS, type=str, default=DEFAULT_YOSYS)
 	p.add_argument('--clock',           help='clock signal name (default: %s)' % DEFAULT_CLK, type=str, default=DEFAULT_CLK)
 	p.add_argument('--timeout',         help='timeout (CPU time) in seconds (default: %s)' % DEFAULT_TIMEOUT, type=int, default=DEFAULT_TIMEOUT)
-	p.add_argument('--memout',          help='memory limit in mega bytes (default: %s)' % DEFAULT_RANDOM, type=int, default=DEFAULT_MEMOUT)
+	p.add_argument('--memout',          help='memory limit in mega bytes (default: %s)' % DEFAULT_MEMOUT, type=int, default=DEFAULT_MEMOUT)
 	p.add_argument('-s', '--split',     help='toggles transforming system by splitting variables at extract points (default: %r)' % DEFAULT_SPLIT, action="count", default=0)
+	p.add_argument('-g', '--granularity',help='abstract granularity level (between 0-2) (default: %r)' % DEFAULT_GRANULARITY, type=int, default=DEFAULT_GRANULARITY)
 	p.add_argument('-r', '--random',    help='toggles using random ordering and random seed (default: %r)' % DEFAULT_RANDOM, action="count", default=0)
 	p.add_argument('-v', '--verbosity', help='verbosity level (default: %r)' % DEFAULT_VERBOSITY, type=int, default=DEFAULT_VERBOSITY)
 	args, leftovers = p.parse_known_args()
@@ -42,7 +44,7 @@ def getopts(header):
 header="""
 	Averroes (avr) -- Abstract VERification of Reachability Of Electronic Systems
 	Versions """ + str(version) + """
-	Reads a Verilog-2005 file and performs property checking with data abstraction.
+	Reads a Verilog file and performs property checking using syntactic data abstraction.
 		supports SystemVerilog concurrent assertions
 
 	Copyright (C) 2018  Aman Goel <amangoel@umich.edu> and Karem A. Sakallah <karem@umich.edu>, University of Michigan
@@ -55,12 +57,11 @@ header="""
 	3. Z3      (Copyright (c) 2015 Microsoft Corporation)
 	
 	---------------------------------
-	Limitiations (as of Aug 01, 2018)
+	Limitiations (as of Oct 25, 2018)
 	---------------------------------
 	1. Can only handle safety properties that can be expressed without temporal operators.
-	2. Assumes single clock domain.
-	3. Cannot handle asynchronous flops.
-	4. Handles memory by simply expanding it using yosys.
+	2. Cannot handle asynchronous flops.
+	3. Handles memory by simply expanding it using Yosys.
 		(customize the bin/avr for special preprocessing using Yosys)
 		
 	Please report bugs via email (amangoel@umich.edu) or on github (https://github.com/aman-goel/avr)
@@ -76,6 +77,7 @@ def split_path(name):
 	return head, tail
 
 def main():
+	print(short_header)
 	known, opts = getopts(header)
 	if not os.path.isfile(opts.bin + "/avr"):
 		raise Exception("avr: main shell script not found")
@@ -86,11 +88,14 @@ def main():
 	if not os.path.isfile(opts.bin + "/reach"):
 		raise Exception("avr: reach binary not found")
 	if not os.path.isfile(opts.yosys + "/yosys"):
-		raise Exception("Please install Yosys using build.sh (or from https://github.com/aman-goel/yosys)")
+		if not os.path.isfile("/usr/local/bin/yosys"):
+			raise Exception("Please install Yosys using build.sh (or from https://github.com/aman-goel/yosys)")
+		else:
+			opts.yosys = "/usr/local/bin"
+			print("\tfound Yosys in /usr/local/bin")
 	if not os.path.isfile(opts.file):
 		raise Exception("Unable to find top file: %s" % opts.file)
 	
-	print(short_header)
 	path, f = split_path(opts.file)
 	command = ""
 	command = command + "./" + opts.bin + "/avr"
@@ -109,6 +114,10 @@ def main():
 		split = not DEFAULT_SPLIT
 	command = command + " " + str(split)
 		
+	command = command + " " + str(opts.granularity)
+	if known.top is not None:
+		command = command + " " + str(opts.top)
+	
 	random = DEFAULT_RANDOM
 	if (opts.random % 2 == 1):
 		random = not DEFAULT_RANDOM
